need test large img trans

+++ src/Core/Worker.cpp
@@ -76,7 +76,7 @@ void Worker::CompleteEvent(EventPackage* event) {
     ProcessEvent(event);
     switch (event->m_eventType) {
         case EVENT_TYPE_WRITE:
-            Debug("6 Write 0x{:X}\n", (char*)(event->ioves[1].iov_base) + event->total_send);
+            Debug("6 Write 0x{:X}\n", (size_t)((char*)(event->ioves[1].iov_base) + event->total_send));
 	        uring.addWrite(event, event->m_fd, (char*)(event->ioves[1].iov_base) + event->total_send, event->ioves[1].iov_len - event->total_send);
             break;
         case EVENT_TYPE_WRITEV:
